Refactor logger library so its constants can be reused elsewhere without triggering its primature initialisation
