+++ sugar/lib/logger/manager.py
@@ -10,6 +10,7 @@ from twisted.python.logfile import LogFile
 
 from sugar.utils.objects import Singleton
 from sugar.lib import six
+from sugar.lib.logger import Logger
 from sugar.config import get_config
 
 
@@ -52,44 +53,6 @@ class SugarLogObserver(log.FileLogObserver):
                                                        ssm=event_data['system'], txt=msg)
             util.untilConcludes(self.write, msg)
             util.untilConcludes(self.flush)
-
-
-class Logger(object):
-    """
-    Turns log levels into `twisted.python.log.msg` corresponding method alias.
-    Result is before:
-
-      log.msg(message, level=logging.INFO, system='hello')
-
-    After:
-
-      log.info(message)
-
-    """
-    LOG_LEVELS = {
-        'all': logging.NOTSET,
-        'debug': logging.DEBUG,
-        'error': logging.ERROR,
-        'critical': logging.CRITICAL,
-        'info': logging.INFO,
-        'warning': logging.WARNING,
-        'fatal': logging.FATAL
-    }
-
-    def __init__(self, name):
-        self.name = name
-        for method in self.LOG_LEVELS:
-            def make_log_level_caller(level):
-                """
-                Wrap Twisted's log into levels
-                :param level:
-                :return:
-                """
-                def _msg(message):
-                    log.msg(message, level=level, system=self.name)
-                return _msg
-
-            setattr(self, method, make_log_level_caller(self.LOG_LEVELS[method]))
 
 
 @Singleton
